move scraper and data
